// src/api/database/global-types.ts
export enum UniversitiesOrderBy {
  CITY_ID_ASC = "CITY_ID_ASC",
  CITY_ID_DESC = "CITY_ID_DESC",
  CREATED_AT_ASC = "CREATED_AT_ASC",
  CREATED_AT_DESC = "CREATED_AT_DESC",
  ID_ASC = "ID_ASC",
  ID_DESC = "ID_DESC",
  LONG_NAME_ASC = "LONG_NAME_ASC",
  LONG_NAME_DESC = "LONG_NAME_DESC",
  NATURAL = "NATURAL",
  PRIMARY_KEY_ASC = "PRIMARY_KEY_ASC",
  PRIMARY_KEY_DESC = "PRIMARY_KEY_DESC",
  SHORT_NAME_ASC = "SHORT_NAME_ASC",
  SHORT_NAME_DESC = "SHORT_NAME_DESC",
  UPDATED_AT_ASC = "UPDATED_AT_ASC",
  UPDATED_AT_DESC = "UPDATED_AT_DESC",
}
// ...
export interface UniversityInput {
  cityId: any;
  shortName: string;
  longName?: string | null;
}