// src/api/database/fragments/types/PartialUniversity.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: PartialUniversity
// ====================================================

export interface PartialUniversity {
  __typename: "University";
  id: any;
  shortName: string;
  longName: string | null;
  createdAt: any;
  updatedAt: any;
  cityId: any;
}